Delete view page, add Mailchimp template submit to Preview page with UI and response feedback
